updated wiremock file

+++ packages/vsce/__tests__/__e2e__/resources/start-wiremock.sh
@@ -1,8 +1,13 @@
+if [ -z "$1" ]; then
+  echo "No directory provided. Please provide a directory for WireMock mappings."
+  exit 1
+fi
+
 docker run -it --rm -d \
     -p 8080:8080 \
     --name wiremock \
     -v "$1"
-    wiremock/wiremock:3.10.0
+    wiremock/wiremock:latest
 
 sleep 1
